services/api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3
pydantic-settings==2.8.1
PyYAML==6.0.2
requests==2.32.3
praw==7.8.1
feedparser==6.0.11
python-dateutil==2.9.0.post0
scikit-learn==1.6.1
numpy==2.0.2
pandas==2.2.3
httpx==0.28.1
pytest==8.3.5
